requirements.txt
asgiref==3.3.1
autopep8==1.5.6
Django==3.1.7
django-cpf==0.1.0
django-crispy-forms==1.11.2
django-stdimage==5.3.0
Pillow==8.1.2
pycodestyle==2.7.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
yapf==0.30.0
